+++ __tests__/01_basic_spec.ts
@@ -1,12 +1,14 @@
 import {
-  LazyFetcher,
+  FetchData,
+  FetchDataLazy,
   useRouteData,
   useRouteDataSelector,
 } from '../src/index';
 
 describe('basic spec', () => {
   it('exported function', () => {
-    expect(LazyFetcher).toBeDefined();
+    expect(FetchData).toBeDefined();
+    expect(FetchDataLazy).toBeDefined();
     expect(useRouteData).toBeDefined();
     expect(useRouteDataSelector).toBeDefined();
   });