+++ examples/03_nesting/src/MyRoutes.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import { Routes, Route, LazyFetcher } from 'react-suspense-router';
+import { Routes, Route, FetchDataLazy } from 'react-suspense-router';
 
 const Index = React.lazy(() => import('./pages/Index'));
 const Posts = React.lazy(() => import('./pages/Posts'));
 const FirstUser = React.lazy(() => import('./pages/FirstUser'));
 const User = React.lazy(() => import('./pages/User'));
-const fetchFirstUserData = LazyFetcher(() => import('./pages/FirstUser.data'));
-const fetchUserData = LazyFetcher(() => import('./pages/User.data'));
+const fetchFirstUserData = FetchDataLazy(() => import('./pages/FirstUser.data'));
+const fetchUserData = FetchDataLazy(() => import('./pages/User.data'));
 
 const MyRoutes: React.FC = () => (
   <Routes>