+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Changed
 - Improve SSR example with Stream
+- Rename LazyFetcher to FetchDataLazy [BREAKING CHANGE]
+### Added
+- New FetchData for async function
 
 ## [1.0.0-alpha.5] - 2020-02-11
 ### Changed